+++ src/C3.Infrastructure/Configuration/ISettingsUpgradeStore.vb
@@ -1,0 +1,15 @@
+Namespace Configuration
+
+Public Interface ISettingsUpgradeStore
+
+    Property UpgradeRequired As Boolean
+
+    Sub UpgradeFromPreviousVersion()
+
+    Sub Normalize()
+
+    Sub Save()
+
+End Interface
+
+End Namespace